Update README.md
The Ephemeral Pages is your gateway to the enchanting world of scrapbook journaling. We create and curate beautiful scrapbooks and provide an array of decorative elements—ribbons, washi tape, stickers, and more—to help you capture and cherish your most treasured memories. Let your creativity flourish with our carefully designed products that add a touch of elegance and personal charm to every page.